backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.2
fastapi-cli==0.0.12
fastapi-cloud-cli==0.1.5
filetype==1.2.0
frozenlist==1.7.0
google-ai-generativelanguage==0.7.0
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.31.1
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-experimental==0.3.4
langchain-google-genai==2.1.12
langchain-groq==0.3.8
langchain-text-splitters==0.3.11
langsmith==0.4.28
markdown-it-py==4.0.0
markupsafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.6.4
mypy-extensions==1.1.0
narwhals==2.5.0
numexpr==2.13.0
numpy==2.3.3
orjson==3.11.3
packaging==25.0
pandas==2.3.2
plotly==6.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.11.0
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
sentry-sdk==2.38.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.43
starlette==0.48.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
typer==0.17.4
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.25.0
